chore: prettier and refactor
