+++ package.json
@@ -23,8 +23,7 @@
     "@agoric/casting": "dev",
     "@agoric/ertp": "^0.14.2",
     "@agoric/ui-components": "^0.2.33",
-    "@agoric/wallet": "^0.13.1",
-    "@agoric/web-components": "0.3.1-dev-8ac3c4c.0",
+    "@agoric/web-components": "^0.5.0",
     "@agoric/zoe": "0.24.0",
     "@endo/eventual-send": "^0.16.0",
     "@endo/init": "0.5.43",
